reorg Position events
